make-rpm.sh: add proof-of-concept script for making rpms
This script uses the internal Red Hat brew service to build an rpm
for a go package based on a git branch. The git branch must contain
a specfile where the `commit` and `shortcommit` globals are unset.

To use the script you must be on the Red Hat VPN and authenitacted using
kerberose aganist REDHAT.COM.

Example:
    $ kinit <you>@REDHAT.COM
    $ ./make-rpm.sh mnt/rpm git@github.com:teg/osbuild-composer.git rhel-8.2.0 golang-github-osbuild-composer.spec rhel-8.2.0 x86_64
This will get the spec file and sources from the given repository. The
specfile will be ammended with the commit information of the head of the
branch. The sources will be ammended by adding vendoring for each of the
go dependencies. A scratch build is then made in brew and downloaded to
`mnt/rpm`.

If a scratch build of the osbuild dependency is also in `mnt/rpm` you can
try the package out:
    $ ./vm.sh rhel-guest-image-8.2-115.x86_64.qcow2
Once booted the rpms are installed, the service is running and both cockpit
and composer-cli are available to try it out.

Signed-off-by: Tom Gundersen <teg@jklm.no>